[][Fix PPPoE WAN->LAN Unbind Issue for MT798x]
[Description]
Fix PPPoE WAN->LAN un-bind issue for MT798x.

The original checking point for multicast packets
is too early, where the target netdev has not yet
been assigned with "real" MAC address. Under this
situation, it might mis-lead HNAT driver not to
bind the traffic that is supposed to be accelerated.
So we move back the checking point to POST_ROUTING
and eventually the issue can be solved.

   $ static unsigned int skb_to_hnat_info(...) {
   $      if (!hnat_priv->data->mcast &&
   $          is_multicast_ether_addr(eth->h_dest))
   $               return 0;
   $ }

If without this patch, PPPoE WAN->LAN traffic cannot
be properly accelerated.

[Release-log]
N/A

Change-Id: I0022631856a0c09d241ab2bc434d07d66803bf27
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/5542217